tweaking the interactions to make it look more accurate

+++ src/com/bbj/cva/screenobjects/IScreenObject.java
@@ -7,8 +7,8 @@ public interface IScreenObject {
 	public void create();
 	public void destroy();
 	public void render(SpriteBatch spriteBatch);
-	public void setX(int x);
-	public void setY(int y);
+	public void setX(float x);
+	public void setY(float y);
 	public float getWidth();
 	public float getHeight();
 	public void loadTextureIfNeeded();
